src/simple_endpoint/requirements.txt
djangorestframework==3.9.4
Django==2.2.4
pygit2==0.28.2